[~] Coding Style

+++ src/client/ui/wchat.c
@@ -16,9 +16,9 @@ static void show_buffer(win_t *this)
 		memcpy(buffer, this->ui->buffer + i * (COLS - 6), COLS - 6);
 		mvwprintw(this->ncurses_win, 2 + i, 4, "%s", buffer);
 	}
-	if (this->ui->buffer_idx <= COLS - 7)
+	if (this->ui->buffer_idx <= (size_t) COLS - 7)
 		mvwprintw(this->ncurses_win, 2, 4 + this->ui->buffer_idx, "|");
-	else if (this->ui->buffer_idx < COLS * 2 - 12)
+	else if (this->ui->buffer_idx < (size_t) COLS * 2 - 12)
 		mvwprintw(this->ncurses_win, 3,
 			this->ui->buffer_idx % (COLS - 6) + 4, "|");
 	free(buffer);
